# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Dating App API",
    "version": "1.0.0",
    "description": "API documentation for the dating app backend"
  },
  "servers": [
    {
      "url": "http://localhost:5000/api/v1",
      "description": "Local server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/admin/users": {
      "get": {
        "summary": "Get admin review list of users",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "type": "string",
              "example": "UNDER_REVIEW"
            },
            "description": "Filter users by status"
          }
        ],
        "responses": {
          "200": {
            "description": "Users fetched successfully"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Admin access required"
          }
        }
      }
    },
    "/admin/users/{id}": {
      "get": {
        "summary": "Get full details of one user for admin review",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "example": "dev-user-2"
            },
            "description": "User ID"
          }
        ],
        "responses": {
          "200": {
            "description": "User fetched successfully"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Admin access required"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/admin/users/{id}/approve": {
      "post": {
        "summary": "Approve a user profile",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "example": "dev-user-2"
            },
            "description": "User ID"
          }
        ],
        "responses": {
          "200": {
            "description": "User approved successfully"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Admin access required"
          }
        }
      }
    },
    "/admin/users/{id}/reject": {
      "post": {
        "summary": "Reject a user profile",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "example": "dev-user-2"
            },
            "description": "User ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": {
                    "type": "string",
                    "example": "Incomplete profile"
                  }
                },
                "required": [
                  "reason"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User rejected successfully"
          },
          "400": {
            "description": "Rejection reason is required"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Admin access required"
          }
        }
      }
    },
...
  },
...
}